+++ crates/dcc-mcp-gateway-admin/Cargo.toml
@@ -15,6 +15,7 @@ workspace = true
 [dependencies]
 axum.workspace = true
 chrono = { version = "0.4", features = ["serde"] }
+dcc-mcp-catalog.workspace = true
 dcc-mcp-db = { path = "../dcc-mcp-db", default-features = false, optional = true }
 dcc-mcp-models.workspace = true
 dcc-mcp-transport.workspace = true
